+++ src/sdk/WalletStorage.interfaces.ts
@@ -189,6 +189,11 @@ export interface FindSincePagedArgs {
   since?: Date
   paged?: Paged
   trx?: TrxToken
+  /**
+   * Support for orderDescending is implemented in StorageKnex for basic table find methods,
+   * excluding certificate_fields table, map tables, and settings (singleton row table).
+   */
+  orderDescending?: boolean
 }
 
 export interface FindForUserSincePagedArgs extends FindSincePagedArgs {