+++ test/Wallet/support/operations.man.test.ts
@@ -1,9 +1,10 @@
-import { WalletOutput } from '@bsv/sdk'
+import { Beef, MerklePath, WalletOutput } from '@bsv/sdk'
 import { sdk, TableOutput, TableUser, verifyOne, verifyOneOrNone } from '../../../src'
-import { _tu } from '../../utils/TestUtilsWalletStorage'
+import { _tu, logger } from '../../utils/TestUtilsWalletStorage'
 import { specOpInvalidChange, ValidListOutputsArgs } from '../../../src/sdk'
 import { LocalWalletTestOptions } from '../../utils/localWalletMethods'
 import { Format } from '../../../src/utility/Format'
+import { asString } from '../../../src/utility/utilityHelpers.noBuffer'
 
 describe('operations.man tests', () => {
   jest.setTimeout(99999999)
@@ -48,13 +49,17 @@ describe('operations.man tests', () => {
 
   test('1 review and unfail false doubleSpends', async () => {
     const { env, storage, services } = await _tu.createMainReviewSetup()
-    let offset = 2700
+    let offset = 0
     const limit = 100
     let allUnfails: number[] = []
     for (;;) {
       let log = ''
       const unfails: number[] = []
-      const reqs = await storage.findProvenTxReqs({ partial: { status: 'doubleSpend' }, paged: { limit, offset } })
+      const reqs = await storage.findProvenTxReqs({
+        partial: { status: 'doubleSpend' },
+        paged: { limit, offset },
+        orderDescending: true
+      })
       for (const req of reqs) {
         const gsr = await services.getStatusForTxids([req.txid])
         if (gsr.results[0].status !== 'unknown') {
@@ -75,13 +80,17 @@ describe('operations.man tests', () => {
 
   test('2 review and unfail false invalids', async () => {
     const { env, storage, services } = await _tu.createMainReviewSetup()
-    let offset = 800
+    let offset = 0
     const limit = 100
     let allUnfails: number[] = []
     for (;;) {
       let log = ''
       const unfails: number[] = []
-      const reqs = await storage.findProvenTxReqs({ partial: { status: 'invalid' }, paged: { limit, offset } })
+      const reqs = await storage.findProvenTxReqs({
+        partial: { status: 'invalid' },
+        paged: { limit, offset },
+        orderDescending: true
+      })
       for (const req of reqs) {
         if (!req.txid || !req.rawTx) continue
         const gsr = await services.getStatusForTxids([req.txid])
@@ -97,6 +106,59 @@ describe('operations.man tests', () => {
     }
     for (const id of allUnfails) {
       await storage.updateProvenTxReq(id, { status: 'unfail' })
+    }
+    await storage.destroy()
+  })
+
+  test.skip('3 review proven_txs', async () => {
+    const { env, storage, services } = await _tu.createMainReviewSetup()
+    let offset = 0
+    const limit = 100
+    let allUnfails: number[] = []
+    //for (const provenTxId of  [3064, 3065, 11268, 11269, 11270, 11271] ) {
+    for (let height = 895000; height < 895026; height++) {
+      let log = ''
+      const unfails: number[] = []
+      const txs = await storage.findProvenTxs({ partial: { height }, paged: { limit, offset } })
+      for (const tx of txs) {
+        const gmpr = await services.getMerklePath(tx.txid)
+        if (gmpr && gmpr.header && gmpr.merklePath) {
+          const mp = gmpr.merklePath
+          const h = gmpr.header
+          const mr = mp.computeRoot(tx.txid)
+          const index = mp.path[0].find(leaf => leaf.hash === tx.txid)?.offset!
+
+          const mp2 = MerklePath.fromBinary(tx.merklePath)
+          const mr2 = mp2.computeRoot()
+
+          if (h.height !== mp.blockHeight || h.merkleRoot !== mr) {
+            console.log(`Merkle root mismatch for ${tx.txid} ${h.merkleRoot} != ${mr}`)
+          } else {
+            if (
+              tx.merkleRoot !== mr ||
+              tx.height !== mp.blockHeight ||
+              tx.blockHash !== h.hash ||
+              tx.index !== index ||
+              mp2.blockHeight !== tx.height ||
+              mr2 !== tx.merkleRoot ||
+              asString(tx.merklePath) !== asString(mp.toBinary())
+            ) {
+              debugger
+              await storage.updateProvenTx(tx.provenTxId, {
+                merklePath: mp.toBinary(),
+                merkleRoot: mr,
+                height: mp.blockHeight,
+                blockHash: h.hash,
+                index
+              })
+              log += `updated ${tx.provenTxId}\n`
+            }
+          }
+        }
+      }
+      //console.log(`${offset} ${log}`)
+      //if (txs.length < limit) break
+      //offset += txs.length
     }
     await storage.destroy()
   })
@@ -172,6 +234,63 @@ describe('operations.man tests', () => {
       status: ['unmined', 'completed'],
       paged: { limit: 100, offset: countReqs - 100 }
     })
+    await storage.destroy()
+  })
+
+  test.skip('12 check storage merged BEEF', async () => {
+    const userId = 127
+    const txid = 'efba8b92a22c3308f432b292b5ec7efb3869ecd50c62cb3ddfb83871bc7be194'
+    const vout = 1
+    const { env, storage, services } = await _tu.createMainReviewSetup()
+
+    const ptx = verifyOne(await storage.findProvenTxs({ partial: { txid } }))
+
+    const mp = MerklePath.fromBinary(ptx.merklePath)
+    expect(mp.blockHeight).toBe(ptx.height)
+
+    const txids = ['24b19a5179c1f146e825643df4c6dc2ba21674828c20fc2948e105cb1ca91eae']
+
+    const r = await storage.getReqsAndBeefToShareWithWorld(txids, [])
+
+    await storage.destroy()
+  })
+
+  test('13 review use of outputs in all following transactions', async () => {
+    const { env, storage, services } = await _tu.createMainReviewSetup()
+
+    const txids = ['2df7b5059112a42fc40adb54ee36244cee0dd216c35ad6c4b6ef4631c14a0e83'] //, '9fb38fc87c6ff39f5c7321a4c689db535c024498ed20031434485c981dd7a182', '3fb6b02e1d001dded1daee3f59dcd684489b96a35a9dfb5082b4119a31689966', '72ea8d84a4c54dbca292f4a79a5ff08cb9917fc3127c1dcff0628aeba8b40823', '0564a515566bc43c1396becf12bbf2d82d821ae7b6e0ef404eedfa090d4877c2', '3b93e4327a50a7f4a421af9fbdec0206b3b7ba5252bc5a0142d0d64aa34c2e73', 'd4b0c3d820696afad43b43e095f3b8c3df52385bb4aeddff0212e0a472dd8e4e']
+    const userId = 111
+    const txs = await storage.findTransactions({
+      partial: { userId },
+      status: ['completed', 'unproven', 'failed'],
+      orderDescending: true,
+      paged: { limit: 50 }
+    })
+    const allTxids = txs.map(tx => tx.txid!)
+    debugger
+    const reqs = await storage.findProvenTxReqs({ partial: {}, txids: allTxids })
+    const beef = new Beef()
+    for (const req of reqs) {
+      beef.mergeRawTx(req.rawTx!)
+    }
+
+    for (const txid of txids) {
+      const o = await storage.findOutputs({ partial: { txid, userId } })
+      const tx = await storage.findTransactions({ partial: { txid, userId } })
+      if (o && tx) {
+        const ltx = await Format.toLogStringTableTransaction(tx[0], storage)
+        logger(ltx)
+        for (const btx of beef.txs) {
+          const tx = btx.tx!
+          for (const i of tx.inputs) {
+            if (i.sourceTXID === txid && i.sourceOutputIndex === 0) {
+              const sltx = Format.toLogStringBeefTxid(beef, btx.txid)
+              logger(sltx)
+            }
+          }
+        }
+      }
+    }
     await storage.destroy()
   })
 })