+++ CHANGELOG.md
@@ -3,6 +3,30 @@
 This document captures the history of significant changes to the wallet-toolbox repository. The git commit history contains the details but is unable to draw
 attention to changes that materially alter behavior or extend functionality.
 
+## wallet-toolbox 1.3.30
+
+- Enable gorillaPoolArc for postBeef Services
+- Switch Services postBeef multi-service mode from 'PromiseAll' to 'UntilSuccess'
+
+## wallet-toolbox 1.3.29
+
+- add verifyUnlockScripts to both createAction and signAction flows
+  
+## wallet-toolbox 1.3.28
+
+- adminStats now includes monitorStats and servicesStats of type ServicesCallHistory (wallet-toolbox/src/sdk/WalletServices.interfaces.ts)
+- both sets of stats break down service calls by providers including both recent calls and interval based statistics.
+- monitorStats correspond to service requests made by the active Monitor daemon. This includes “delayed” createActions. Intervals are currently 12 minutes.
+- servicesStats corresponds to the service requests made by the StorageProvider service. This includes “non-delayed” createActions. Intervals are determined by rate of calls to adminStats, each call starts a new interval.
+
+## wallet-toolbox 1.3.25
+
+- throws INVALID_PARAMETER if a createAction input is a change output.
+- logging and potential fix for internalizeAction bug.
+- adds gorillaPool to Services but leaves it disabled for now.
+- adds service call history logging to Monitor Events table, but not yet tied in to adminStats return value.
+- StorageProvider level “find” entity methods now support additional optional orderDescending boolean.
+
 ## wallet-toolbox v1.3.4, 2025-04-24
 
 ### Add StorageIdb